+++ lib/lsp.dart
@@ -1,3 +1,4 @@
 export 'src/protocol/language_server/interface.dart' show LanguageServer;
 export 'src/protocol/language_server/messages.dart';
 export 'src/protocol/language_server/server.dart' show StdIOLanguageServer;
+export 'src/protocol/language_server/wireformat.dart' show lspChannel;